requirements.txt
pyrekordbox
pytesseract
pyautogui
pygetwindow
Pillow
